% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sas.R
\name{convert_sas}
\alias{convert_sas}
\title{Convert sas7bdat file to either rds or csv format}
\usage{
convert_sas(file, write_path, to_ext = "rds")
}
\arguments{
\item{file}{the name of the file which the data are to be read from.
If it does not contain an absolute path, the file name is relative to the
current working directory.}

\item{write_path}{the filepath to save the converted file. The directory will
be created if not already present.}

\item{to_ext}{either 'csv' or 'rds' options are available.}
}
\description{
Reads a file in sas7bdat format and converts to either a comma separated
value or an R serialized data format
}
