% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sas_dir.R
\name{convert_sas_dir}
\alias{convert_sas_dir}
\title{Convert  multiple sas7bdat files}
\usage{
convert_sas_dir(file_path, write_path, to_ext = "rds")
}
\arguments{
\item{file_path}{the name of the directory which the data are to be read from.
If it does not contain an absolute path, the directory is relative to the
current working directory.}

\item{write_path}{the filepath to save the converted file. The directory will
be created if not already present.}

\item{to_ext}{either 'csv' or 'rds' options are available.}
}
\description{
Reads multiple files from a directory in sas7bdat format and converts to
either a comma separated value or an R serialized data format
}
